refactor: util_decorator로 이름 변경 적용
