+++ app/api/controllers/album_duplicate_controller.py
@@ -3,7 +3,7 @@ from app.core.cache import get_cached_embedding, get_cached_embeddings_parallel
 from fastapi.responses import JSONResponse
 from app.schemas.album_schema import ImageRequest
 from app.service.duplicate import find_duplicate_groups
-from app.utils.logging_utils import log_exception
+from app.utils.logging_decorator import log_exception
 
 
 @log_exception
