+++ app/api/controllers/album_embedding_controller.py
@@ -1,7 +1,7 @@
 from app.utils.image_loader import get_image_loader
 from app.service.embedding import embed_images  # service import
 from app.schemas.album_schema import ImageRequest
-from app.utils.logging_utils import log_exception
+from app.utils.logging_decorator import log_exception
 from fastapi import Request
