+++ app/api/endpoints/album_category_router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 from app.api.controllers.album_category_controller import categorize_controller
 from app.schemas.album_schema import ImageRequest
-from app.utils.logging_utils import log_flow
+from app.utils.logging_decorator import log_flow
 from fastapi import Request
 
 router = APIRouter(tags=["category"])
