+++ app/api/endpoints/album_duplicate_router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 from app.api.controllers.album_duplicate_controller import duplicate_controller
 from app.schemas.album_schema import ImageRequest
-from app.utils.logging_utils import log_flow
+from app.utils.logging_decorator import log_flow
 
 router = APIRouter(tags=["duplicate"])
