+++ app/api/endpoints/album_score_router.py
@@ -3,7 +3,7 @@ from app.api.controllers.album_score_controller import (
     highlight_scoring_controller,
 )
 from app.schemas.album_schema import CategoryScoreRequest
-from app.utils.logging_utils import log_flow
+from app.utils.logging_decorator import log_flow
 
 router = APIRouter(tags=["score"])
