+++ app/service/embedding.py
@@ -1,6 +1,6 @@
 import torch
 from app.core.cache import set_cached_embedding
-from app.utils.logging_utils import log_exception
+from app.utils.logging_decorator import log_exception
 
 @log_exception
 def embed_images(
